requirements.txt
fastapi
uvicorn[standard]
aiohttp
opencv-python
numpy
python-multipart
pillow